feat : 게시판 글 목록 구현

+++ app/src/main/java/com/example/logisticsestimate/data/BoardData.kt
@@ -4,8 +4,8 @@ import com.google.gson.annotations.SerializedName
 
 data class BoardData(
     @SerializedName("category") var category: Int,
-    @SerializedName("postId") var postId: Int,
-    @SerializedName("uid") var uid: Int,
+    @SerializedName("postId") var postId: Long,
+    @SerializedName("uid") var uid: Long,
     @SerializedName("date") var date: String,
 
     @SerializedName("postTitle") var postTitle: String,
